requirements.txt
aiida-core==0.12.2
ejplugins>=0.9.7
packaging==19.0
spglib==1.10.3.65
ase==3.12.0
# testing
mock==2.0.0
pgtest==1.1.0
sqlalchemy-diff==0.1.3
pytest==3.6.3
wheel>=0.31
coverage
pytest-cov
pytest-timeout
jsonextended